[HatohlDBUtils] Merge information about trigger.
Until now, some members in EventInfo such as severity and brief
have been filled using a join operation with the trigger table,
when it's retrieved. This way has the following two demerits.

(1) It's hard to delete records in the trigger table.
(2) The number of the save of an event is only once, however
    it may be read many times with join operations.

So this patch resolves the problems by storing duplicated information
in the event table.